bugfix and cleanup
  * Bugfix: X-Prometheus-Scrape-Timeout-Seconds is a string
    representation of a float value.  int() cannot parse it,
    but float() can.
  * Removed requirements.txt as it is no longer needed.

+++ prometheus_swagger_exporter/__init__.py
@@ -98,7 +98,7 @@ def get_url(target, request_path, spec_segment):
 
 @route('/probe')
 def metrics():
-    timeout = int(request.headers.get(
+    timeout = float(request.headers.get(
         'X-Prometheus-Scrape-Timeout-Seconds',
         default=CheckerBase.nrpe_timeout
     ))
